Simplified and bugfixed changes from DvO chunk9 and in general
in certreq_new:
intentionally removing the checks in certreq_new which would not lead to a failure in this fuction:
- if srvCert or trusted_store is not available, it will be detected during protection verification if it would be an issue.
- if there is no mean (referenceNo, clCert) to protect, there could be a reason,
  if that's accidentially it will come to the surface anyway during message protection or when the server
  rejects the message

Moved error handling back to CMP_*_new messages, removing one degree of refactoring. This is in order
not to hand over OpenSSL error defines greated with myerr.pl - which has potential to eventually lead to issues.

For KUR, the subject name must be copied also in case of subjectAltNames being available.
Changed that accordingly and also set from oldClCert if available.

Removed the following comment from CMP_kur_new, as there is now a differentiation between clCert and the optional oldClCert.
    * TODO: the differentiation between certificate used to sign the CMP messages
    * and the certificate to update should be improved - so far only the clCert
    * could be updated

Amended docu for CMP_CTX_subjectAltName_push1() to highlight there that setting subjectAltName
will prevent the clCert subject name to be copied for IR and CR.