small improvements over old code in cmp_ses.c which got refactored in DvO's chunk 7a
Turned around the logic when checking return value when polling for a response
in cert_response() - to make it more logical (good case first).

Removed comment regarding possibility to use caPubs directly as TA,
as it is not needed at this particular place. However storing info
about the way the message was encrypted was created as feature request in
sourceforge.