Renaming the user/pass CLI options to ref/secret
to be alligned to the language in RFC 4210 and to avoid confusion with other
"xyz-pass" options related to private keys.

Note: the RFC does sometimes use "Reference Number" while it's not clear whether
the reference actually must be a number; the KeyIdentifier is an ASN.1 octet
string.  So it was aimed to have the docu language open in that respect.