Made it bit more clear that there is no explicit "RevReqID" in the standard,
but rather that it is the (implicit ASN.1) sequence id which must match in
the RevReq and RevRep.

Moved #defines for first CertReqID and REVREQ(S)ID from cmp.h to cmp_int.h as
there should be no need to utilize those #defines from outside the library
code.

Removed training whitespaces in CMP_lib.pod