Revert "fix issue #131, due to subtle build issue for shared libs depending on .a extension"
This reverts commit eeb1a12212b472df557274871ec67a9aa6a513bf.

This broke building the test with shared libraries enabled.  If the DEPEND in
build.info is not for libcrypto.a but just for libcrypto, OpenSSL scripting
attempts to use the shared library during linking.  libcrypto.so however is
cleaned from functions which are not exported, but the tests depend on internal
functions.