Moving senderKID and extraCerts checks to be after failed message validation
No need to check that and log any errors in case the message validation did not
fail.  Warnings on all ways how the server certificate candidate was determined
as it does not matter if the certificate was e.g. given as srvCert.

Warning on server cert not in extraCerts
- also if there is no extraCerts as that might be the cause of the error
- on all ways how a certificate was given.  Especially if senderKID is missing,
  the server cert finding algorithm might not have found the particular needed
  certificate which might be in extraCerts - but another one e.g. from untrusted
  certs store first.

Note that RFC 5280 does not require but, only recommend to have senderKID in EE
certificates (RA certs might not be CA certs).  In most cases this would likely
be rather stupid, but any way permitted.