Transformed fail if senderKID does not match given srvCert to a warning
There's actually no requirement in the RFC to fail in case senderKID would not
be correct, it only says that "senderKID ... usable to indicate".  Anyway,
message validation will hopefully fail if the key is wrong.

Made it a warning to help debugging e.g. in case a server uses a wrong key than
it claims and it is desired to test that by manually setting srvCert.