First shot on the TLS-via-proxy feature
Very strangely, that appears not to receive the first HTTP/1.0 200 from the proxy
which acks that the connection is established.  BIO_gets returns -1, so it's not
even a partial receive.

Eventually, for some strange reason, it sends the CONNECT strings out a 2nd
time with TCP PSH flag set.  As this is received by the CMP server behind the
proxy, a HTTP/1.0 400 is received - which actualy gets received.

I don't manage to figure out why this happens. :(