Certificate Management Protocol (CMP, RFC 4210) extension
Also includes HTTP transfer (RFC 6712)

CMP and CRMF API is added to libcrypto, and the "cmp" app to the openssl CLI.
Adds extensive man pages and unit tests.  Integration into build scripts.

Squashed version of the CMP patch as of 2019-03-20
 last incremental commit 09d7d67f7f455197c2a72b115b4234bc95110628

This builds up on the CRMF (RFC 4211) code which is in OpenSSL master, following
pull request 7646 (and 8527).